requirements.txt
Flask
reportlab
arabic_reshaper
python-bidi
